Don't reseed before test calls
This patch changes the reseeding behavior to only reseed at test
setup/teardown, but not again when running the test. This means that the
state will no longer be reset between a fixture and the test using the
fixture, which enables tests like this to work as expected:

```python
import random
import pytest

@pytest.fixture
def random_number():
    return random.random()

def test_something(random_number):
    new_number = random.random()
    assert new_number != random_number
```

This example is simple, but becomes important when fixtures use
Factory-Boy to generate more complex random data.