Add an entry point to register external random engines (#187)
This adds the entry point `pytest_randomly.random_seeder`. Anything specified under this entry point will be called with the static seed value every time the random generators are reseeded, as though it were one of the `.seed()` functions.

The current method of adding large packages directly is great, but we've got a package that has it's own random generator and isn't nearly big enough to be worth integrating into this package. This patch solves that by allowing any external packages to register an entry point for pytest-randomly. If pytest-randomly isn't installed then the extra entry point is harmless.

This is a single-function interface at the moment so means that you can't re-set the state - it has to be generated from the seed every time. I don't know if there was some reason other than speed that the generators are set from state if possible, rather than seed. If this is a problem, there could be an object API to allow getting/setting state but unless we need the complication I didn't want to go that far in an initial proposal.

Since I'm not sure the canonical way to test `pkg_resources`, I've put a test in that monkeypatches the interface. Depending on your taste/rigour this might not be acceptable.

This would also presumably add an avenue for fixing #135 and any other packages that want similar.